test: Convert default explain limit tests to new explain setup
